fixed internal error handling and moved implementation to header file
