+++ libc/src/__support/HashTable/randomness.h
@@ -14,7 +14,7 @@
 #include "src/__support/macros/attributes.h"
 #include "src/__support/macros/config.h"
 #if defined(LIBC_HASHTABLE_USE_GETRANDOM)
-#include "src/__support/OSUtil/getrandom.h"
+#include "src/__support/OSUtil/linux/getrandom.h"
 #include "src/__support/libc_errno.h"
 #endif
 
@@ -39,14 +39,14 @@ LIBC_INLINE uint64_t next_random_seed() {
     size_t count = sizeof(entropy);
     uint8_t *buffer = reinterpret_cast<uint8_t *>(entropy);
     while (count > 0) {
-      ssize_t len = internal::getrandom(buffer, count, 0);
-      if (len == -1) {
+      auto len = internal::getrandom(buffer, count, 0);
+      if (!len.has_value()) {
         if (libc_errno == ENOSYS)
           break;
         continue;
       }
-      count -= len;
-      buffer += len;
+      count -= len.value();
+      buffer += len.value();
     }
     libc_errno = errno_backup;
 #endif
