--- libc/src/__support/OSUtil/getrandom.h
@@ -1,23 +1,0 @@
-//===------------ Implementation of getrandom function ----------*- C++ -*-===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
-#ifndef LLVM_LIBC_SRC___SUPPORT_OSUTIL_GETRANDOM_H
-#define LLVM_LIBC_SRC___SUPPORT_OSUTIL_GETRANDOM_H
-
-#include "src/__support/macros/config.h"
-#include <sys/random.h>
-
-namespace LIBC_NAMESPACE_DECL {
-namespace internal {
-
-ssize_t getrandom(void *buf, size_t buflen, unsigned int flags);
-
-} // namespace internal
-} // namespace LIBC_NAMESPACE_DECL
-
-#endif // LLVM_LIBC_SRC___SUPPORT_OSUTIL_GETRANDOM_H
