+++ libc/src/__support/OSUtil/linux/getrandom.h
@@ -1,4 +1,4 @@
-//===------------ Linux implementation of getrandom -------------*- C++ -*-===//
+//===------------ Implementation of getrandom function ----------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,25 +6,30 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "src/__support/OSUtil/getrandom.h"
+#ifndef LLVM_LIBC_SRC___SUPPORT_OSUTIL_GETRANDOM_H
+#define LLVM_LIBC_SRC___SUPPORT_OSUTIL_GETRANDOM_H
+
+#include "hdr/types/ssize_t.h"
 #include "src/__support/OSUtil/linux/syscall.h" // syscall_impl
 #include "src/__support/common.h"
-#include "src/__support/libc_errno.h"
+#include "src/__support/error_or.h"
 #include "src/__support/macros/config.h"
-#include <sys/syscall.h> // For syscall numbers.
+#include <sys/syscall.h> // For syscall numbers
 
 namespace LIBC_NAMESPACE_DECL {
 namespace internal {
 
-ssize_t getrandom(void *buf, size_t buflen, unsigned int flags) {
+static inline ErrorOr<ssize_t> getrandom(void *buf, size_t buflen,
+                                         unsigned int flags) {
   ssize_t ret =
       LIBC_NAMESPACE::syscall_impl<ssize_t>(SYS_getrandom, buf, buflen, flags);
   if (ret < 0) {
-    libc_errno = static_cast<int>(-ret);
-    return -1;
+    return Error(-static_cast<int>(ret));
   }
   return ret;
 }
 
 } // namespace internal
 } // namespace LIBC_NAMESPACE_DECL
+
+#endif // LLVM_LIBC_SRC___SUPPORT_OSUTIL_GETRANDOM_H
