+++ libc/src/sys/random/linux/getrandom.cpp
@@ -8,9 +8,10 @@
 
 #include "src/sys/random/getrandom.h"
 
-#include "src/__support/OSUtil/getrandom.h"
+#include "src/__support/OSUtil/linux/getrandom.h"
 #include "src/__support/OSUtil/syscall.h" // For internal syscall function.
 #include "src/__support/common.h"
+#include "src/__support/error_or.h"
 
 #include "src/__support/libc_errno.h"
 #include "src/__support/macros/config.h"
@@ -20,7 +21,11 @@ namespace LIBC_NAMESPACE_DECL {
 
 LLVM_LIBC_FUNCTION(ssize_t, getrandom,
                    (void *buf, size_t buflen, unsigned int flags)) {
-  return internal::getrandom(buf, buflen, flags);
+  auto rand = internal::getrandom(buf, buflen, flags);
+  if (rand.has_value())
+    return rand.value();
+  libc_errno = static_cast<int>(-(rand.error()));
+  return -1;
 }
 
 } // namespace LIBC_NAMESPACE_DECL
